Some refactoring of our hashing code.
Generally I'm looking at wrapping some of the crypto primitives in
easier to use abstractions. I've started with the hashing algorithms by
refactoring and adding tests for one of the SHA hashes we've
implemented. I plan to add more hash functions that comply with the
Hasher interface in future commits.

I've also added a Hashable interface that will eventually be used for
digital signatures, I.E. simply implement Hashable for a struct and it
can be signed.