feat(signing): new signing algorithm and related cleanup
Added a curve25519 edwards signing algorithm and tests. Also removed the
Hashable interface. This was originally meant to help with deterministic
hashing of protos, but we've solved this in another way and it simply
adds complication now.

Breaks anywhere hashing/signing is done (using hashable). We should be
updating to the ed25519 anyways which this will be a part of.